requirements.txt
jax==0.5.0
flax==0.10.2
optax==0.2.4
diffrax==0.6.2
hydra-core==1.3.2
omegaconf==2.3.0